Add ente Auth - Free / Cross Platform / E2EE 2FA app
Hey @pluja. I took the liberty of moving this as the first item in the list, as it is the only one from the list that is cross platform, and it matches (or is better than) the existing entries in all other aspects. But I might be thinking biased (since I'm one of the developers), so please feel free to reorder.

Some mentions/reviews:
- [New oil - Twitter mention](https://twitter.com/thenewoil1/status/1602077944611995649)
- [Detailed review at zeroclick.xyz](https://zeroclick.xyz/ente-auth)
- [FOSS Android - Mastodon review](https://mstdn.social/@foss_android/109505185418700364)

And here is the original blog post introducing it - https://ente.io/blog/auth/

Thank you!